Move prompt converters to a separate module. Camelcase local variables and add missing JSDocs.
